+++ src/components/index.js
@@ -80,6 +80,7 @@ import MatchInfos from './match/MatchInfos'
 import MatchTeamStats from './match/MatchTeamStats'
 
 // Profile
+import AddSocialModal from './profile/AddSocialModal'
 import FavoriteWeaponCard from './profile/FavoriteWeaponCard'
 import HeatmapStatsCard from './profile/HeatmapStatsCard'
 import LevelStatsCard from './profile/LevelStatsCard'
@@ -89,6 +90,8 @@ import MatchHistoryPaginationItem from './profile/MatchHistoryPaginationItem'
 import MatchHistoryStatsLink from './profile/MatchHistoryStatsLink'
 import ProfileCard from './profile/ProfileCard'
 import ProfileHeader from './profile/ProfileHeader'
+import ProfileHeaderButtons from './profile/ProfileHeaderButtons'
+import ProfileHeaderSocialButtons from './profile/ProfileHeaderSocialButtons'
 
 // Notifications
 import NotificationList from './notifications/NotificationList'
@@ -130,6 +133,7 @@ import UserMenuOptions from './menu/UserMenuOptions'
 export {
   AcceptIcon,
   AccountCard,
+  AddSocialModal,
   AddUserIcon,
   ArrowDownIcon,
   ArrowRightIcon,
@@ -199,6 +203,8 @@ export {
   PodiumIcon,
   ProfileCard,
   ProfileHeader,
+  ProfileHeaderButtons,
+  ProfileHeaderSocialButtons,
   Progress,
   RefuseIcon,
   RequireAuth,