+++ src/hooks/useProfileDetails.jsx
@@ -1,0 +1,9 @@
+import { useContext } from 'react'
+
+import { ProfileDetailsContext } from '@contexts'
+
+const useProfileDetails = () => {
+  return useContext(ProfileDetailsContext)
+}
+
+export default useProfileDetails